+++ Dog/src/br/com/dog/model/entity/Match.java
@@ -1,40 +1,47 @@
 package br.com.dog.model.entity;
 
 public class Match {
-	private int idMatch;
-	private int requisitante;
-	private int requisitado;
+	private long idMatch;
+	private long requisitante;
+	private long requisitado;
 	private String start;
-	
-	public int getIdMatch() {
+
+	public long getIdMatch() {
 		return idMatch;
 	}
-	public void setIdMatch(int idMatch) {
+
+	public void setIdMatch(long idMatch) {
 		this.idMatch = idMatch;
 	}
-	public int getRequisitante() {
+
+	public long getRequisitante() {
 		return requisitante;
 	}
-	public void setRequisitante(int requisitante) {
+
+	public void setRequisitante(long requisitante) {
 		this.requisitante = requisitante;
 	}
-	public int getRequisitado() {
+
+	public long getRequisitado() {
 		return requisitado;
 	}
-	public void setRequisitado(int requisitado) {
+
+	public void setRequisitado(long requisitado) {
 		this.requisitado = requisitado;
 	}
+
 	public String getStart() {
 		return start;
 	}
+
 	public void setStart(String start) {
 		this.start = start;
 	}
+
 	@Override
 	public String toString() {
-		return "Match [idMatch=" + idMatch + ", requisitante=" + requisitante
-				+ ", requisitado=" + requisitado + ", start=" + start + "]";
+		return "Match [idMatch=" + idMatch + ", requisitante=" + requisitante + ", requisitado=" + requisitado
+				+ ", start=" + start + "]";
 	}
-	
 
 }
